pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "django-caps"
description = "Capability based object permissions for Django."
version = "0.0.1"
authors = ["Thomas Baquet <thomas oxylus app>"]
readme = "README.rst"
license = "MPL-2.0"
packages = [
    { include = "caps" },
]
include = [
    {path="docs", format="sdist"},
    {path="tests", format="sdist"},
    {path="instance", format="sdist"},
]
exclude = [
]

classifiers = [
    "Framework :: Django",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
]


[tool.poetry.dependencies]
python = "^3.13"
Django = "^5"
djangorestframework = "^3"
django-filter = "^25.2"


[tool.poetry.group.dev.dependencies]
ipython = "^8.29.0"
pre-commit = "^3.1"
black = "^23.3"
ruff = "^0.0"

pytest = "^8"
pytest-django = "^4"
pytest-asyncio = "^0.25.3"

[tool.poetry.group.docs.dependencies]
Sphinx = "^8"
sphinx-rtd-theme = "^3"
sphinx-autodoc-typehints = "^3"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "instance.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]


[tool.black]
line-length = 120
exclude = '''
    /(
          \.egg
        | \.git
        | \.hg
        | \.tox
        | \._build
        | \.build
        | \.bulk-out
        | \.dist
        | \.__pycache__
        | \.venv
        | \.migrations
        | \.static
        | \.ox/settings
        | \.scripts
    )
'''


[tool.ruff]
line-length = 120
exclude = [
    "egg",
    "git",
    "hg",
    "tox",
    "_build",
    "build",
    "dist",
    "__pycache__",
    "venv",
    "*/migrations",
    "*/static",
    "scripts/",
]


[tool.semantic_release]
version_variables = ["caps/__init__.py:__version__"]
version_toml = [
    "pyproject.toml:tool.poetry.version",
]
